+++ test/types/index.ts
@@ -2970,3 +2970,17 @@ type a = keyof {}
 			}>()
 		})
 }
+
+// ? Promise<Response>
+{
+	async function handler() {
+		return new Response(JSON.stringify({ text: 'hello' }), {
+			status: 200,
+			headers: { 'Content-Type': 'application/json' }
+		})
+	}
+
+	new Elysia().get('/hello', () => handler(), {
+		response: { 200: t.Object({ text: t.String() }) }
+	})
+}